Fix signedness warning in gio/giomodule.c
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from gobject/gbinding.h:28,
                 from glib/glib-object.h:23,
                 from gio/gioenums.h:28,
                 from gio/giotypes.h:28,
                 from gio/giomodule.h:28,
                 from gio/giomodule.c:25:
gio/giomodule.c: In function ‘print_help’:
glib/gmacros.h:348:26: error: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Werror=sign-compare]
 #define MAX(a, b)  (((a) > (b)) ? (a) : (b))
                          ^
gio/giomodule.c:720:19: note: in expansion of macro ‘MAX’
           width = MAX (width, strlen (g_io_extension_get_name (extension)));
                   ^~~
glib/gmacros.h:348:35: error: operand of ?: changes signedness from ‘int’ to ‘size_t’ {aka ‘long unsigned int’} due to unsignedness of other operand [-Werror=sign-compare]
 #define MAX(a, b)  (((a) > (b)) ? (a) : (b))
                                   ^~~
gio/giomodule.c:720:19: note: in expansion of macro ‘MAX’
           width = MAX (width, strlen (g_io_extension_get_name (extension)));
                   ^~~

+++ gio/giomodule.c
@@ -712,7 +712,7 @@ print_help (const char        *envvar,
     {
       GList *l;
       GIOExtension *extension;
-      int width = 0;
+      gsize width = 0;
 
       for (l = g_io_extension_point_get_extensions (ep); l; l = l->next)
         {
